feat/backup : Edit CursorPageResponseBackupDto(...Short size) -> int size

+++ src/main/java/com/team2/hrbank/backup/dto/CursorPageResponseBackupDto.java
@@ -6,7 +6,7 @@ public record CursorPageResponseBackupDto(
         List<BackupDto> content,
         String nextCursor,
         Long nextIdAfter,
-        Short size,
+        int size,
         Long totalElements,
         Boolean hasNext
 ) {
